rename nipxx to nip111
