+++ .gitignore
@@ -5,4 +5,4 @@ yarn-error.log
 package-lock.json
 .envrc
 test.html
-!.env.nipxx
+!.env.nip111
