+++ index.ts
@@ -11,7 +11,7 @@ export * as nip19 from './nip19'
 export * as nip26 from './nip26'
 export * as nip39 from './nip39'
 export * as nip57 from './nip57'
-export * as nipxx from './nipxx'
+export * as nip111 from './nip111'
 
 export * as fj from './fakejson'
 export * as utils from './utils'
