+++ lib/esm/nostr.mjs
@@ -1167,9 +1167,9 @@ function makeZapReceipt({
   return zap;
 }
 
-// nipxx.ts
-var nipxx_exports = {};
-__export(nipxx_exports, {
+// nip111.ts
+var nip111_exports = {};
+__export(nip111_exports, {
   loginWithX: () => loginWithX,
   privateKeyFromX: () => privateKeyFromX,
   registerWithX: () => registerWithX,
@@ -1259,11 +1259,11 @@ export {
   nip04_exports as nip04,
   nip05_exports as nip05,
   nip06_exports as nip06,
+  nip111_exports as nip111,
   nip19_exports as nip19,
   nip26_exports as nip26,
   nip39_exports as nip39,
   nip57_exports as nip57,
-  nipxx_exports as nipxx,
   relayInit,
   serializeEvent,
   signEvent,
