+++ lib/nostr.cjs.js
@@ -39,11 +39,11 @@ __export(nostr_tools_exports, {
   nip04: () => nip04_exports,
   nip05: () => nip05_exports,
   nip06: () => nip06_exports,
+  nip111: () => nip111_exports,
   nip19: () => nip19_exports,
   nip26: () => nip26_exports,
   nip39: () => nip39_exports,
   nip57: () => nip57_exports,
-  nipxx: () => nipxx_exports,
   relayInit: () => relayInit,
   serializeEvent: () => serializeEvent,
   signEvent: () => signEvent,
@@ -1212,9 +1212,9 @@ function makeZapReceipt({
   return zap;
 }
 
-// nipxx.ts
-var nipxx_exports = {};
-__export(nipxx_exports, {
+// nip111.ts
+var nip111_exports = {};
+__export(nip111_exports, {
   loginWithX: () => loginWithX,
   privateKeyFromX: () => privateKeyFromX,
   registerWithX: () => registerWithX,
