+++ nip111.test.js
@@ -1,17 +1,17 @@
 /* eslint-env jest */
 const ethers = require('ethers')
 const fetch = require('node-fetch')
-const { nipxx, nip05 } = require('./lib/nostr.cjs')
+const { nip111, nip05 } = require('./lib/nostr.cjs')
 globalThis.crypto = require('crypto')
-require('dotenv').config({ path: '.env.nipxx' })
+require('dotenv').config({ path: '.env.nip111' })
 
 privateKey = process.env.ETHER_PRIVATE_KEY;
 let wallet = new ethers.Wallet(privateKey);
 
-nipxx.useFetchImplementation(fetch)
+nip111.useFetchImplementation(fetch)
 nip05.useFetchImplementation(fetch)
 
-// @dev: ↓ YOU MUST ENTER THE CORRESPONDING PRIVATE KEY IN THE .env.nipxx FILE
+// @dev: ↓ YOU MUST ENTER THE CORRESPONDING PRIVATE KEY IN THE .env.nip111 FILE
 let address = '0x0d2C290bb3fE24D8D566268d5c41527c519715Db' // ← signing checksummed ethereum pubkey/address
 let info = `eip155:1:${address}`
 
@@ -22,17 +22,17 @@ test('Private Key from Deterministic Signature and Identifiers', async () => {
   let signature = 'f'.padEnd(130, 'f')
   let password = 'hello dear fucker'
   expect(
-    await nipxx.privateKeyFromX(username, info, signature, password)
+    await nip111.privateKeyFromX(username, info, signature, password)
   ).toEqual('fd6a6c03eadf0db178f79de3a2dd3f0464fb5fac96608842d68ce64da2e40954')
 
   // without password
   expect(
-    await nipxx.privateKeyFromX(username, info, signature)
+    await nip111.privateKeyFromX(username, info, signature)
   ).toEqual('897c5140a6e0e09b512c755cfd60829998fb0d046c52fae0846cca045185a52b')
 
   // 0x+hex signature; SHOULD BE agnostic to '0x' prefix
   expect(
-    await nipxx.privateKeyFromX(username, info, '0x' + signature, password)
+    await nip111.privateKeyFromX(username, info, '0x' + signature, password)
   ).toEqual('fd6a6c03eadf0db178f79de3a2dd3f0464fb5fac96608842d68ce64da2e40954')
 })
 
@@ -43,7 +43,7 @@ test('Login with Deterministic Signature and NIP-02 Identifiers', async () => {
   let signature = 'f'.padEnd(130, 'f')
   let password = 'hello dear fucker'
   expect(
-    await nipxx.signInWithX(username, info, signature, password)
+    await nip111.signInWithX(username, info, signature, password)
   ).toEqual({
     'petname': username,
     'privkey': '3f0b9c0ddb02d7056c4908d362c8d03072ef3f2dc1eb10a6ef9706dc3f017198',
@@ -61,7 +61,7 @@ test('Generate Public Key for NIP-05 Identifier without verifying record', async
   let signature = await promise;
   let password = ''
   expect(
-    await nipxx.signInWithXStandalone('zuck@cash.app', info, signature, password)
+    await nip111.signInWithXStandalone('zuck@cash.app', info, signature, password)
   ).toEqual({
     'petname': username.split('@')[0],
     'pubkey': '84ef21c0150c3a6abaf9da9c6e078281c2c1af09abb72a6518dce2ff82c9cb45',
@@ -75,7 +75,7 @@ test('Invalid Signature/Password in NIP-05 Identifiers', async () => {
   let signature = 'f'.padEnd(130, 'f')
   let password = ''
   await expect(async () => {
-    await nipxx.signInWithX(username, info, signature, password)
+    await nip111.signInWithX(username, info, signature, password)
   }).rejects.toThrow('Invalid Signature/Password');
 })
 
@@ -85,7 +85,7 @@ test('NIP-05 Identifiers Not Set', async () => {
   let signature = 'f'.padEnd(130, 'f')
   let password = 'hello dear fucker'
   await expect(async () => {
-    await nipxx.signInWithX('zuck@cash.app', info, signature, password)
+    await nip111.signInWithX('zuck@cash.app', info, signature, password)
   }).rejects.toThrow('Nostr Profile Not Found');
 })
 
@@ -99,7 +99,7 @@ test('Login with Deterministic Signature and NIP-05 Identifiers', async () => {
   let password = 'hello dear fucker'
 
   expect(
-    await nipxx.signInWithX(username, info, signature, password)
+    await nip111.signInWithX(username, info, signature, password)
   ).toEqual({
     'petname': username.split('@')[0],
     'profile': {
